routers/workflow.py
import uuid
from typing import Dict, Any, Optional, List
from fastapi import APIRouter, HTTPException, Depends
from pydantic import BaseModel

from workflows.prd_agent import app_graph, checkpointer
from dependencies import logger

router = APIRouter(prefix="/api/workflow", tags=["workflow"])

class WorkflowStartRequest(BaseModel):
    prd_content: str
    improvement_type: str = "comprehensive"

class FeedbackRequest(BaseModel):
    feedback: Optional[str] = None
    approve: bool = True

class ChatRequest(BaseModel):
    messages: List[Dict[str, str]]
    prd_context: Optional[str] = ""

@router.post("/chat")
async def chat_with_assistant(request: ChatRequest):
    """Context-aware AI assistant for PRD refinement."""
    from dependencies import get_llm_client, logger
    
    client = get_llm_client()
    if not client:
        return {"success": False, "error": "AI features are currently disabled."}
        
    if not request.messages:
         return {"success": False, "error": "No messages provided"}

    try:
        # Construct system prompt with PRD context
        system_message = {
            "role": "system",
            "content": (
                "You are an expert Product Manager assistant helping to refine a Product Requirements Document (PRD).\n"
                "Your goal is to provide specific, actionable advice or content for the PRD.\n"
                "When asked to write or improve a section, keep it consistent with the existing tone.\n"
                "CURRENT PRD CONTEXT:\n"
                f"\"\"\"\n{request.prd_context}\n\"\"\"\n\n"
                "If the user asks for a new section, provide it in clear Markdown format."
            )
        }
        
        # Prepare full message list for LLM
        llm_messages = [system_message] + request.messages
        
        # Use simple client call since this is stateless chat (not the agent workflow)
        # Note: We hardcode model for now or get from env
        response = client.chat.completions.create(
            model="Azure-GPT-5-chat", # Configurable
            messages=llm_messages,
            temperature=0.7
        )
        
        reply = response.choices[0].message.content
        return {"success": True, "reply": reply}
        
    except Exception as e:
        logger.error(f"Chat error: {str(e)}")
        return {"success": False, "error": str(e)}

@router.post("/start")
async def start_workflow(request: WorkflowStartRequest):
    """Start a new document improvement workflow."""
    logger.info("Received workflow start request")
    thread_id = str(uuid.uuid4())
    config = {"configurable": {"thread_id": thread_id}}
    
    initial_state = {
        "prd_content": request.prd_content,
        "improvement_type": request.improvement_type,
        "messages": []
    }
    
    try:
        logger.info(f"Invoking graph for thread {thread_id}...")
        result = await app_graph.ainvoke(initial_state, config=config)
        logger.info(f"Graph invocation returned for thread {thread_id}")
        
        # Get the next state to see where we are
        state_snapshot = await app_graph.aget_state(config)
        logger.info(f"Current state next: {state_snapshot.next}")
        
        return {
            "thread_id": thread_id,
            "status": "waiting_for_review",
            "next_step": state_snapshot.next,
            "current_content": state_snapshot.values.get("improved_content"),
            "original_content": request.prd_content
        }
    except Exception as e:
        logger.error(f"Workflow start error: {e}", exc_info=True)
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/{thread_id}/state")
async def get_workflow_state(thread_id: str):
    config = {"configurable": {"thread_id": thread_id}}
    state_snapshot = await app_graph.aget_state(config)
    
    if not state_snapshot:
        raise HTTPException(status_code=404, detail="Thread not found")
        
    return {
        "thread_id": thread_id,
        "status": "active" if state_snapshot.next else "completed",
        "next_step": state_snapshot.next,
        "current_content": state_snapshot.values.get("improved_content")
    }

@router.post("/{thread_id}/review")
async def review_workflow(thread_id: str, request: FeedbackRequest):
    """Provide human review/feedback."""
    config = {"configurable": {"thread_id": thread_id}}
    state_snapshot = await app_graph.aget_state(config)
    
    if not state_snapshot.next:
        raise HTTPException(status_code=400, detail="Workflow already completed or not active")
    
    if request.approve:
        # Proceed to end (human_review -> END)
        # We can update state if needed, but for now just resume.
        # Command(resume=None) is implicit if we just invoke again?
        # Typically we use update_state or just invoke with None input to resume.
        
        # In this simple graph, resuming from 'human_review' just goes to END.
        result = await app_graph.ainvoke(None, config=config)
        return {
            "thread_id": thread_id,
            "status": "completed",
            "final_content": result.get("improved_content")
        }
    else:
        # If rejected/feedback, we want to go BACK to 'generate'.
        # We can update state with feedback and force the next node to be 'generate'.
        
        await app_graph.aupdate_state(
            config, 
            {"feedback": request.feedback}, 
            as_node="human_review" # Simulate being in human review
        )
        # Now forcing redirection content logic is tricky in linear graph without conditional edges.
        # Ideally the graph definition should have a conditional edge from 'human_review':
        #   if Approved -> END
        #   if Feedback -> generate
        
        # Since I defined linear generate->human_review->END, I need to modify the graph definition 
        # to support the loop back. But "update_state" allows "time travel" effectively or state modification.
        # Actually, best practice is conditional edge.
        # But this is "implementation" phase, so I should probably fix the graph definition first.
        # For now, I will error if feedback is provided saying "Feedback loop not implemented in POC graph yet"
        # OR I can just say "Approved" finishes it.
        
        # Let's fix the graph definition in the next step or assume approval for the POC.
        # I'll stick to simple "Approved" path for now as per plan "User (via API) will approve or edit the content."
        # If they *edit*, they likely replace the content.
        
        # If request has new content (edits), we validly save that as final.
        return {"status": "rejected", "message": "Feedback loop requires graph update."}

