requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.0
openai>=1.55.0
langgraph>=0.1.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
itsdangerous==2.1.2
starsessions==1.3.0
python-docx==1.1.0
PyPDF2==3.0.1
pymupdf==1.23.8
tenacity==8.2.3
markdown==3.5.1
authlib==1.3.0
httpx==0.26.0